feat: user management module done with full functionallity.
security feature in the auth module fixed.
UserManagementController and UserManagementService added to auth module.
app errors refactored.
feign implemented for send requests from user-management to auth.

+++ sale/src/main/java/dev/faruk/sale/dto/SalePostRequest.java
@@ -12,10 +12,10 @@ import java.util.List;
 @AllArgsConstructor
 public class SalePostRequest {
     private Double receivedMoney;
-    private Integer cashierId;
+    private Long cashierId;
     private List<ProductDetails> products;
 
-    public SalePostRequest(Double receivedMoney, Integer cashierId) {
+    public SalePostRequest(Double receivedMoney, Long cashierId) {
         this.receivedMoney = receivedMoney;
         this.cashierId = cashierId;
     }
